Improve code organization
