+++ users/api/transport.go
@@ -18,8 +18,9 @@ import (
 )
 
 // MakeHandler returns a HTTP handler for Users and Groups API endpoints.
-func MakeHandler(cls users.Service, authn smqauthn.AuthNMiddleware, tokensvc grpcTokenV1.TokenServiceClient, selfRegister bool, mux *chi.Mux, logger *slog.Logger, instanceID string, pr *regexp.Regexp, idp supermq.IDProvider, providers ...oauth2.Provider) http.Handler {
-	mux = usersHandler(cls, authn, tokensvc, selfRegister, mux, logger, pr, idp, providers...)
+func MakeHandler(svc users.Service, authn smqauthn.AuthNMiddleware, tokensvc grpcTokenV1.TokenServiceClient, selfRegister bool, mux *chi.Mux, logger *slog.Logger, instanceID string, pr *regexp.Regexp, idp supermq.IDProvider, providers ...oauth2.Provider) http.Handler {
+	mux = usersHandler(svc, authn, tokensvc, selfRegister, mux, logger, pr, idp, providers...)
+	mux = oauthHandler(mux, svc, tokensvc, providers...)
 
 	mux.Get("/health", supermq.Health("users", instanceID))
 	mux.Handle("/metrics", promhttp.Handler())
